requirements.txt
matplotlib.pyplot
numpy
tkinter